Refactor sample api spec
